ai-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
scikit-learn
joblib
groq
langchain-huggingface
langchain-community
sentence-transformers
pypdf